+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod builder;
 pub mod cli;
 pub mod client;
+pub mod commons;
 pub mod crd;
 pub mod error;
 pub mod label_selector;